[Chore] combine `timer.ts` into `async.ts`
